deps/crypto: Add hash algorithms
Split the algorithms code out of the AA eventlog and move it to the local
`crypto` crate. Doing this allows other crates to consume the code rather
than duplicating the logic.
