+++ attestation-agent/deps/crypto/src/lib.rs
@@ -31,3 +31,6 @@ pub use symmetric::*;
 
 mod asymmetric;
 pub use asymmetric::*;
+
+mod algorithms;
+pub use algorithms::*;
